refactor: split swiftioc.py into an installable package
Break the 2510-line single-file script into swiftioc/ (models, fp,
extract, http_client, scoring, parsers, collect, writers,
logging_utils, cli), re-exported from swiftioc/__init__.py so
`import swiftioc as si` keeps its existing public surface unchanged.

parsers.py resolves http_get/load_feedparser through a self-import
(`import swiftioc as _pkg`) rather than importing them directly, so
existing `monkeypatch.setattr(swiftioc, "http_get", ...)` calls in the
test suite keep working — a direct import would bind the name at
import time and never see the patched value.

cli.py now sets http_client._SAVE_RAW_DIR / http_client.HTTP_DEBUG as
qualified module-attribute assignments instead of module-local
globals, since those names are read from http_client.py, not cli.py,
after the split.

Adds pyproject.toml packaging metadata (console script, dynamic deps
from requirements.txt) and updates CI/collect workflows plus
README/CONTRIBUTING to invoke `python -m swiftioc`.
